Refactor grammar mistakes and folders name
