+++ src/app/admin/components/DeleteUserModal.jsx
@@ -24,8 +24,6 @@ const Container = styled('div')(({ theme }) => ({
   height: '100vh',
   width: '100vw',
   position: 'fixed',
-  top: 0,
-  left: 0,
   [`& .${classes.modalContainer}`]: {
     width: '40vw',
     '@media (max-width: 768px)': {
