+++ src/app/admin/components/utils.js
@@ -1,15 +1,16 @@
 import * as Yup from 'yup'
 
 export const getAddUserValidationSchema = () => Yup.object({
-  name: Yup.string().required('Nombre requerido'),
-  username: Yup.string().required('Username requerido'),
-  password: Yup.string().min(6, 'Mínimo 6 caracteres').required('Password requerido'),
-  role: Yup.string().oneOf(['ADMIN', 'CASHIER', 'WAREHOUSE']).required('Rol requerido'),
+  name: Yup.string().required('Name is Required'),
+  username: Yup.string().required('Username is Required'),
+
+  password: Yup.string().min(6, 'Minimum 6 characters"').required('Password required'),
+  role: Yup.string().oneOf(['ADMIN', 'CASHIER', 'WAREHOUSE']).required('Role required'),
 })
 
 export const getEditUserValidationSchema = () => Yup.object({
-  name: Yup.string().required('Field required'),
-  username: Yup.string().required('Field required'),
+  name: Yup.string().required('Field is required'),
+  username: Yup.string().required('Field is equired'),
   role: Yup.string().oneOf(['ADMIN', 'CASHIER', 'WAREHOUSE']).required('Field required'),
   password: Yup.string().min(6, 'Minimum 6 characters').optional(),
 })
